Rename utils -> util
Per style.
